git_status: Fix non-null arithmetic exit status
Move from post-increment to pre-increment in the counter implementation
in order to avoid the behaviour where
    (( X++ ))
returns 1 if X evaluates to 0, which is interpreted as a failure exit
status, in turn causing the whole script to terminate prematurely.

Note that, for X non-negative,
    (( ++X ))
never evaluates to 0, avoiding the unintended behaviour.